Fixed BFQ_Simulation not starting
