+++ quaos/pauli.py
@@ -2,6 +2,7 @@ import numpy as np
 import itertools
 import re
 import math 
+import scipy
     
 
 # PAULIS
@@ -213,3 +214,19 @@ def tensor(mm):
         return mm[0]
     else:
         return scipy.sparse.kron(mm[0],tensor(mm[1:]),format="csr")
+
+
+# function for creating generalized Pauli matrix
+def XZ_mat(d,aX,aZ):
+    omega = math.e**(2*math.pi*1j/d)
+    aa0 = np.array([1 for i in range(d)])
+    aa1 = np.array([i for i in range(d)])
+    aa2 = np.array([(i-aX)%d for i in range(d)])
+    X = scipy.sparse.csr_matrix((aa0,(aa1,aa2)))
+    aa0 = np.array([omega**(i*aZ) for i in range(d)])
+    aa1 = np.array([i for i in range(d)])
+    aa2 = np.array([i for i in range(d)])
+    Z = scipy.sparse.csr_matrix((aa0,(aa1,aa2)))
+    if (d == 2) and (aX%2 == 1) and (aZ%2 == 1):
+        return 1j*(X@Z)
+    return X@Z
